Added return codes

+++ app/main.py
@@ -2,7 +2,8 @@ import pytz
 import models
 
 import datetime
-from fastapi import FastAPI, Depends, BackgroundTasks, Request
+from fastapi import FastAPI, Depends, BackgroundTasks, Request, status
+from fastapi.responses import JSONResponse
 from fastapi.templating import Jinja2Templates
 from fastapi.staticfiles import StaticFiles
 
@@ -31,7 +32,7 @@ def get_db():
         db.close()
 
 
-@app.get("/")
+@app.get("/", status_code=status.HTTP_200_OK)
 async def home(request: Request, amount: str = 0, meal_id: str = None, db: Session = Depends(get_db)):
     entries = db.query(GlucoseLevel).order_by(GlucoseLevel.date.desc())
     meals = db.query(Meal)
@@ -47,10 +48,7 @@ async def home(request: Request, amount: str = 0, meal_id: str = None, db: Sessi
 
     for i in entries.all():
 
-        date = i.date.replace(tzinfo=pytz.utc).astimezone(pytz.timezone('America/New_York')).strftime("%Y-%m-%d")
-        # date = date.replace(tzinfo=pytz.utc)
-        # date = date.astimezone(pytz.timezone('America/New_York'))
-        i.date = date
+        i.date = i.date.replace(tzinfo=pytz.utc).astimezone(pytz.timezone('America/New_York')).strftime("%Y-%m-%d")
         format_entries.append(i)
         
 
@@ -64,7 +62,7 @@ async def home(request: Request, amount: str = 0, meal_id: str = None, db: Sessi
     })
 
 
-@app.get("/meal")
+@app.get("/meal", status_code=status.HTTP_200_OK)
 def meal(request: Request, db: Session = Depends(get_db)):
     meals = db.query(Meal).order_by(Meal.name)
 
@@ -73,28 +71,6 @@ def meal(request: Request, db: Session = Depends(get_db)):
         "meals": meals.all()
     })
 
-
-@app.get("/api/glucose/{id}")
-async def get_single_glucose(id: str, db: Session = Depends(get_db)):
-    entry = db.query(GlucoseLevel).filter(GlucoseLevel.id == id).first()
-
-    return entry
-
-@app.delete("/api/glucose/{id}")
-async def delete_glucose_entry(id: str, db: Session = Depends(get_db)):
-    entry = db.query(GlucoseLevel).filter(GlucoseLevel.id == id).first()
-    entry_time = entry.date
-    entry_amount = entry.amount
-    meal_name = entry.meal_id
-    db.delete(entry)
-    db.commit()
-
-    return {
-        "code": "success",
-        "message": f"Deleted ID= {id} MealName={meal_name} Amount={entry_amount} EntryTime={entry_time}"
-    }
-
-
 @app.get("/api/glucose")
 async def get_all_glucose_entry(request: Request, db: Session = Depends(get_db)):
     entries = db.query(GlucoseLevel)
@@ -102,69 +78,90 @@ async def get_all_glucose_entry(request: Request, db: Session = Depends(get_db))
     return entries.all()
 
 
-@app.post("/api/glucose")
-async def add_glucose_entry(glucose_entry: GlucoseEntry, background_tasks: BackgroundTasks, db: Session = Depends(get_db)):
+@app.post("/api/glucose", status_code=status.HTTP_201_CREATED)
+async def add_glucose_entry(glucose_entry: GlucoseEntry, db: Session = Depends(get_db)):
+    try:
+        glucose = GlucoseLevel()
+        glucose.user_id = glucose_entry.user_id
+        glucose.date = glucose_entry.date
+        glucose.amount = glucose_entry.amount
+        glucose.unit = glucose_entry.unit
+        glucose.note = glucose_entry.note
+        glucose.meal_id = glucose_entry.meal_id
 
-    glucose = GlucoseLevel()
-    glucose.user_id = glucose_entry.user_id
-    glucose.date = glucose_entry.date
-    glucose.amount = glucose_entry.amount
-    glucose.unit = glucose_entry.unit
-    glucose.note = glucose_entry.note
-    glucose.meal_id = glucose_entry.meal_id
+        db.add(glucose)
+        db.commit()
 
-    db.add(glucose)
-    db.commit()
-	
-    return {
-	"code": "success",
-	"message": "Entry was added successfully",
-    "id": glucose.id
+        return JSONResponse(
+            status_code=status.HTTP_201_CREATED, 
+            content={
+                'id': glucose.id, 
+                'user_id': glucose.user_id,
+                'meal_id' 
+                'amount': glucose.amount,
+                'message': 'success'})
 
-   }
+    except Exception as e:
+        return JSONResponse(
+            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR, 
+            content={
+                'message': f'Error: {e}'})
 
-@app.get("/api/meal")
+@app.get("/api/glucose/{id}")
+async def get_single_glucose(id: int, db: Session = Depends(get_db)):
+    entry = db.query(GlucoseLevel).filter(GlucoseLevel.id == id).first()
+
+    return entry
+
+@app.delete("/api/glucose/{id}", status_code=status.HTTP_200_OK)
+async def delete_glucose_entry(id: int, db: Session = Depends(get_db)):
+    entry = db.query(GlucoseLevel).filter(GlucoseLevel.id == id).first()
+    
+    try:
+        if entry:
+            entry_time = entry.date
+            entry_amount = entry.amount
+            meal_name = entry.meal_id
+            db.delete(entry)
+            db.commit()
+
+            status_code = status.HTTP_200_OK
+            content = {'id': id, 'message': f'success: deleted {id}'}
+        else:
+            status_code = status.HTTP_404_NOT_FOUND
+            content = {'id': id, 'message': f'error: not found in database {id}'}
+    except Exception as e:
+        status_code = status.HTTP_500_INTERNAL_SERVER_ERROR
+        content = {'message': f'Error: {e}'}
+    finally:
+        return JSONResponse(status_code=status_code, content=content)
+        
+
+@app.get("/api/meal", status_code=status.HTTP_200_OK)
 async def get_meals(request: Request, db: Session = Depends(get_db)):
     entries = db.query(Meal)
 
     return entries.all()
 
-@app.get("/api/meal/{id}")
-async def get_single_meal(id: str, db: Session = Depends(get_db)):
-    entry = db.query(Meal).filter(Meal.id == id).first()
-
-    return entry
-
-@app.delete("/api/meal/{id}")
-async def remove_meal(id: str, db: Session = Depends(get_db)):
-    entry = db.query(Meal).filter(Meal.id == id).first()
-    meal_name = entry.name
-    db.delete(entry)
-    db.commit()
-
-    return {
-        "code": "success",
-        "message": f"Deleted ID= {id} Name={meal_name}"
-    }
-
-
-@app.post("/api/meal")
-async def add_meal(meal: MealEntry, background_tasks: BackgroundTasks, db: Session = Depends(get_db)):
+@app.post("/api/meal", status_code=status.HTTP_201_CREATED)
+async def add_meal(meal: MealEntry, db: Session = Depends(get_db)):
 
     meal_ = Meal()
     meal_.name = meal.name
 
     db.add(meal_)
     db.commit()
-	
-    return {
-	"code": "success",
-	"message": "Entry was added successfully",
-    "id": meal_.id
-   }
 
-@app.put("/api/meal{id}")
-async def add_meal(meal: MealEntry, background_tasks: BackgroundTasks, db: Session = Depends(get_db)):
+    return JSONResponse(status_code=status.HTTP_201_CREATED, content={'id': meal_.id, 'name': meal_.name})
+
+@app.get("/api/meal/{id}")
+async def get_single_meal(id: int, db: Session = Depends(get_db)):
+    entry = db.query(Meal).filter(Meal.id == id).first()
+
+    return entry
+
+@app.put("/api/meal/{id}", status_code=status.HTTP_200_OK)
+async def add_meal(meal: MealEntry, db: Session = Depends(get_db)):
     entry = db.query(Meal).filter(Meal.id == id).first()
 
     meal_ = Meal()
@@ -179,44 +176,73 @@ async def add_meal(meal: MealEntry, background_tasks: BackgroundTasks, db: Sessi
 
    }
 
+@app.delete("/api/meal/{id}", status_code=status.HTTP_200_OK)
+async def remove_meal(id: int, db: Session = Depends(get_db)):
+    entry = db.query(Meal).filter(Meal.id == id).first()
+    if entry:
+        meal_name = entry.name
+        db.delete(entry)
+        db.commit()
+
+        return JSONResponse(
+            status_code=status.HTTP_200_OK, 
+            content={
+                'message': 'Created successfully', 
+                'id': id, 
+                'name': meal_name})
+        
+    else:
+        return JSONResponse(
+            status_code=status.HTTP_404_NOT_FOUND, 
+            content={
+                'id': id, 
+                'message': 'Error: ID not found in database.'
+                })
+
+
 @app.get("/api/user")
 async def get_user(request: Request, db: Session = Depends(get_db)):
     entries = db.query(User)
 
     return entries.all()
 
-@app.get("/api/user/{id}")
-async def get_single_user(id: str, db: Session = Depends(get_db)):
-    entry = db.query(User).filter(User.id == id).first()
+@app.post("/api/user", status_code=status.HTTP_201_CREATED)
+async def add_user(user: UserModel, db: Session = Depends(get_db)):
 
-    return entry
+    # Normalizing the names
+    first_name = user.first_name.lower()
+    last_name = user.last_name.lower()
 
-@app.post("/api/user")
-async def add_user(user: UserModel, background_tasks: BackgroundTasks, db: Session = Depends(get_db)):
+    entry = db.query(User).filter(User.first_name == first_name, User.last_name == last_name).first()
+    if entry:
+        status_code = status.HTTP_409_CONFLICT
+        content = {'message': f"Error: User {first_name} {last_name} already exists in database", 'id': entry.id}
+    else:
+        user_ = User()
+        user_.first_name = user.first_name.lower()
+        user_.last_name = user.last_name.lower()
+        user_.email = user.email.lower()
+        db.add(user_)
+        db.commit()
 
-    user_ = User()
-    user_.first_name = user.first_name.lower()
-    user_.last_name = user.last_name.lower()
-    user_.email = user.email.lower()
+        status_code = status.HTTP_201_CREATED 
+        content = {'id': user_.id, 'message': f"Success: User {first_name} {last_name} added"}
 
-
-    db.add(user_)
-    db.commit()
+    return JSONResponse(status_code=status_code, content=content)
 	
-    return {
-	"code": "success",
-	"message": "Entry was added successfully",
-    "id": user_.id
-
-   }
-
-@app.delete("/api/user/{id}")
-async def remove_user(id: str, db: Session = Depends(get_db)):
+@app.delete("/api/user/{id}", status_code=status.HTTP_200_OK)
+async def remove_user(id: int, db: Session = Depends(get_db)):
     entry = db.query(User).filter(User.id == id).first()
-    db.delete(entry)
-    db.commit()
+    if entry:
+        first_name = entry.first_name
+        last_name = entry.last_name
+        db.delete(entry)
+        db.commit()
 
-    return {
-        "code": "success",
-        "message": f"Deleted ID= {id}"
-    }
+        status_code = status.HTTP_200_OK
+        content = {'id': id, 'message': f"Success: User {first_name} {last_name} deleted"}
+    else:
+        status_code = status.HTTP_404_NOT_FOUND
+        content = {'id': id, 'message': f"Error: User ID={id} not found in database."}
+        
+    return JSONResponse(status_code=status_code, content=content)
